long arithmetic and easter interface

+++ src/main/java/net/lerariemann/infinity/access/NetherPortalBlockAccess.java
@@ -5,5 +5,5 @@ import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 
 public interface NetherPortalBlockAccess {
-    void modifyPortal(World world, BlockPos pos, BlockState state, int i);
+    void modifyPortal(World world, BlockPos pos, BlockState state, long i);
 }
